Save Calendar Modal: Refactor to tsx and use Modal
